// packages/client/src/store/auth/actionsTypes.ts
export enum AuthActionTypes {
  Ready = 'AUTH.READY',
  SetTokens = 'AUTH.SET_TOKENS',
  SetState = 'AUTH.SET_STATE',
  Login = 'AUTH.LOGIN',
  Logout = 'AUTH.LOGOUT',
  Refresh = 'AUTH.REFRESH',
  SignUp = 'AUTH.SIGNUP',
  SetLogin = 'AUTH.SET_LOGIN',
  SetSignUp = 'AUTH.SET_SIGN_UP',
  SetUser = 'AUTH.SET_USER',
  SetRanks = 'AUTH.SET_RANKS',
  SelectCurrency = 'AUTH.SELECT_CURRENCY',
  SetUI = 'AUTH.SET_UI',
  SetOAuth = 'AUTH.SET_OAUTH',
  RestorePassword = 'AUTH.RESTORE_PASSWORD',
  LoginMetamask = 'AUTH.LOGIN_METAMASK',
  SignUpMetamask = 'AUTH.SIGNUP_METAMASK',
  OAuthLogin = 'AUTH.OAUTH_LOGIN',
  OAuthSignup = 'AUTH.OAUTH_SIGNUP',
  LoginTelegram = 'AUTH.LOGIN_TELEGRAM',
  SignUpTelegram = 'AUTH.SING_UP_TELEGRAM',
  AuthSetTelegram = 'AUTH.SET_TELEGRAM',
}